requirements.txt
Keras==2.4.3
numpy==1.22.2
Pillow==8.2.0
streamlit==0.68.0
tenacity==8.0.1
tensorflow==2.5.0rc3